+++ ft_itoa_base.c
@@ -6,15 +6,15 @@
 /*   By: achepurn <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/04 12:51:02 by achepurn          #+#    #+#             */
-/*   Updated: 2017/11/18 21:26:50 by achepurn         ###   ########.fr       */
+/*   Updated: 2017/11/20 18:20:33 by achepurn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	a_len(int n, int base)
+static size_t	a_len(int n, int base)
 {
-	int	res;
+	size_t	res;
 
 	if (n == 0)
 		return (2);
@@ -27,7 +27,7 @@ static int	a_len(int n, int base)
 	return (res);
 }
 
-static void	put_a(int n, char *str, int len, int base)
+static void		put_a(int n, char *str, int len, int base)
 {
 	char	*map;
 
@@ -43,10 +43,10 @@ static void	put_a(int n, char *str, int len, int base)
 	str[len] = map[n];
 }
 
-char		*ft_itoa_base(int n, int base)
+char			*ft_itoa_base(int n, int base)
 {
 	char	*res;
-	int		len;
+	size_t	len;
 
 	if (base == 10)
 		return (ft_itoa(n));
