+++ ft_strnew.c
@@ -6,7 +6,7 @@
 /*   By: achepurn <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/10/26 20:36:05 by achepurn          #+#    #+#             */
-/*   Updated: 2017/11/18 21:54:15 by achepurn         ###   ########.fr       */
+/*   Updated: 2017/11/20 11:52:04 by achepurn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strnew(size_t size)
 {
 	char			*str;
-	unsigned int	i;
+	size_t			i;
 
 	str = (char *)malloc(size + 1);
 	if (str)
